Adds abstract speech visitor for Japanese.
