+++ src/main/java/com/progressiveaccess/cmlspeech/speech/ja/JaSimpleSpeechVisitor.java
@@ -28,13 +28,6 @@
 package com.progressiveaccess.cmlspeech.speech.ja;
 
 import com.progressiveaccess.cmlspeech.analysis.RichStructureHelper;
-import com.progressiveaccess.cmlspeech.connection.Bridge;
-import com.progressiveaccess.cmlspeech.connection.BridgeAtom;
-import com.progressiveaccess.cmlspeech.connection.ConnectingBond;
-import com.progressiveaccess.cmlspeech.connection.SharedAtom;
-import com.progressiveaccess.cmlspeech.connection.SharedBond;
-import com.progressiveaccess.cmlspeech.connection.SpiroAtom;
-import com.progressiveaccess.cmlspeech.speech.AbstractSpeechVisitor;
 import com.progressiveaccess.cmlspeech.speech.Language;
 import com.progressiveaccess.cmlspeech.structure.RichAliphaticChain;
 import com.progressiveaccess.cmlspeech.structure.RichAtom;
@@ -45,51 +38,15 @@ import com.progressiveaccess.cmlspeech.structure.RichFunctionalGroup;
 import com.progressiveaccess.cmlspeech.structure.RichFusedRing;
 import com.progressiveaccess.cmlspeech.structure.RichIsolatedRing;
 import com.progressiveaccess.cmlspeech.structure.RichMolecule;
-import com.progressiveaccess.cmlspeech.structure.RichSetType;
 import com.progressiveaccess.cmlspeech.structure.RichSubRing;
 
-import com.google.common.base.Joiner;
-
 import java.util.Iterator;
-import java.util.SortedSet;
-import java.util.TreeSet;
 
 /**
  * Produces the simple speech for structures.
  */
 
-public class JaSimpleSpeechVisitor extends AbstractSpeechVisitor {
-
-  private boolean shortDescription = false;
-  private boolean subject = true;
-
-
-  @Override
-  public void visit(final RichBond bond) {
-    this.addSpeech(Language.getBondTable().order(bond));
-    this.addSpeech("結合"); // bond
-  }
-
-
-  @Override
-  public void visit(final RichAtom atom) {
-    Integer position = this.getContextPositions().getPosition(atom.getId());
-    // TODO (sorge) Maybe take the supersystem of the atom outside the context.
-    if (position == null) {
-      this.describeSuperSystem(atom);
-      return;
-    }
-    this.addSpeech(Language.getAtomTable().lookup(atom));
-    this.addSpeech(position);
-    if (this.subject) {
-      this.addSpeech("は、"); // Separator (only after subject).
-    }
-    if (this.shortDescription) {
-      return;
-    }
-    this.describeHydrogenBonds(atom);
-  }
-
+public class JaSimpleSpeechVisitor extends JaSpeechVisitor {
 
   @Override
   public void visit(final RichIsolatedRing ring) {
@@ -182,83 +139,6 @@ public class JaSimpleSpeechVisitor extends AbstractSpeechVisitor {
   }
 
 
-  @Override
-  public void visit(final SpiroAtom spiroAtom) {
-    this.shortDescription = true;
-    RichStructureHelper.getRichAtom(spiroAtom.getConnector()).accept(this);
-    this.addSpeech("スピロ原子"); // spiro atom
-    this.addSpeech("に"); // to 
-    RichStructureHelper.getRichAtomSet(spiroAtom.getConnected()).accept(this);
-    this.addSpeech("、"); // Punctuation
-    this.shortDescription = false;
-  }
-
-
-  @Override
-  public void visit(final BridgeAtom bridgeAtom) {
-    RichStructureHelper.getRichAtom(bridgeAtom.getConnector()).accept(this);
-    this.addSpeech("橋頭原子");  // bridge atom
-  }
-
-
-  @Override
-  public void visit(final ConnectingBond bond) {
-    this.shortDescription = true;
-    this.subject = false;
-    String connected = bond.getConnected();
-    if (RichStructureHelper.isAtom(connected)) {
-      RichStructureHelper.getRichAtom(connected).accept(this);
-    } else {
-      RichStructureHelper.getRichAtomSet(connected).accept(this);
-    }
-    this.addSpeech("に"); // to 
-    RichStructureHelper.getRichBond(bond.getConnector()).accept(this);
-    this.addSpeech("、"); // Punctuation
-    // TODO (sorge) The past tense here is problematic!
-    // this.modSpeech("して"); // ed (modifier)
-    this.shortDescription = false;
-    this.subject = true;
-  }
-
-
-  @Override
-  public void visit(final SharedAtom sharedAtom) {
-    this.shortDescription = true;
-    RichStructureHelper.getRichAtom(sharedAtom.getConnector()).accept(this);
-    this.addSpeech("共有原子"); // shared atom
-    RichStructureHelper.getRichAtomSet(sharedAtom.getConnected()).accept(this);
-    this.addSpeech("含有"); // with 
-    this.addSpeech("、"); // Punctuation
-    this.shortDescription = false;
-  }
-
-
-  @Override
-  public void visit(final SharedBond sharedBond) {
-    RichStructureHelper.getRichBond(sharedBond.getConnector()).accept(this);
-    this.addSpeech("共有"); // shared
-  }
-
-
-  @Override
-  public void visit(final Bridge bridge) {
-    this.shortDescription = true;
-    this.addSpeech("fused with");
-    RichStructureHelper.getRichAtomSet(bridge.getConnected()).accept(this);
-    this.addSpeech("via");
-    bridge.getBridges().forEach(c -> c.accept(this));
-    this.shortDescription = false;
-  }
-
-
-  public String getSpeech() {
-    final Joiner joiner = Joiner.on("");
-    String result = joiner.join(this.retrieveSpeech());
-    this.clearSpeech();
-    return result;
-  }
-
-
   // TODO (sorge) For the following utility functions, see if they can be
   // refactored with walk methods, etc.
   private void describeReplacements(final RichAtomSet system) {
@@ -303,70 +183,6 @@ public class JaSimpleSpeechVisitor extends AbstractSpeechVisitor {
       bond.accept(this);
       this.addSpeech("、"); // Punctuation
     }
-  }
-
-
-  private void describeSubstitutions(final RichAtomSet system) {
-    final SortedSet<Integer> subst = new TreeSet<Integer>();
-    for (final String atom : system.getConnectingAtoms()) {
-      subst.add(system.getPosition(atom));
-    }
-    switch (subst.size()) {
-      case 0:
-        return;
-      case 1:
-      default:
-        for (final Integer position : subst) {
-          this.addSpeech(position);
-          this.addSpeech("位"); // position
-          this.addSpeech("と"); // and
-        }
-        this.remSpeech();
-        this.addSpeech("で"); // at
-        this.addSpeech("置換"); // Substitution
-        this.addSpeech("、"); // Punctuation
-        return;
-    }
-  }
-
-
-  /** 
-   * Adds description of hydrogen bonds of an atom.
-   * 
-   * @param atom
-   *          The atom to describe.
-   */
-  private void describeHydrogenBonds(RichAtom atom) {
-    final Integer count = atom.getStructure().getImplicitHydrogenCount();
-    switch (count) {
-      case 0:
-        return;
-      case 1:
-      default:
-        this.addSpeech("水素");  // hydrogen (and hydrogens)
-        this.addSpeech(count.toString());
-        // this.addSpeech("に結合しており、"); // bonded to
-        this.addSpeech("に結合、"); // bonded to
-        return;
-    }
-  }
-
-
-  private void describeSuperSystem(final RichAtom atom) {
-    this.shortDescription = true;
-    for (String context : atom.getContexts()) {
-      if (RichStructureHelper.isAtomSet(context)) {
-        RichAtomSet set = RichStructureHelper.getRichAtomSet(context);
-        RichSetType type = set.getType();
-        if (type == RichSetType.FUNCGROUP
-            || type == RichSetType.ISOLATED
-            || type == RichSetType.FUSED
-            || type == RichSetType.ALIPHATIC) {
-          set.accept(this);
-        }
-      }
-    }
-    this.shortDescription = false;
   }
 
 }
